Don't check SSL certificate (#17).
Indeed, `moh-data` (and `pandas`) calls `urlopen()` and this does SSL verfication. However, for some reasons, this doesn't work from within OpenCOR (while it's fine from within `virtualenv`). So, here, we are going for a drastic approach which is not to check SSL certification in *all* cases. This is usually highly discouraged, but... no other solution at this stage.